qemu.mk: fixes/improvements for Travis CI environment
- If $(CROSS_COMPILE) is defined, use it to set the other cross compile
variables used in common.mk: CROSS_COMPILE_{S,NS}_{USER,KERNEL}.
The linux and qemu targets are also adjusted to use those variables,
they may be updated later when common targets are added.
- Use $(MAKE) instead of plain 'make' so that some special make flags
are handled properly (-n for instance).
- Fix filelist-tee dependencies. Since this target uses 'find' to make
a list of files built by xtest, xtest is a dependency.
- make check: add DUMP_LOGS_ON_ERROR variable. To dump serial0.log and
serial1.log when test fails, run:
    make check DUMP_LOGS_ON_ERROR=1
- Add missing + before $(MAKE). The $(bios-qemu-common) macro defines
a command that must start with a '+' character for reasons related to
the GNU make jobserver implementation [1]. Without it, the sub-make
would print the following warning and not run in parallel:
    make[1]: warning: jobserver unavailable: using -j1.  Add `+' to
    parent make rule.
  [1] http://make.mad-scientist.net/papers/jobserver-implementation/
  ("The Final Result", point 7.)
- Remove redundant dependencies

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>