qemu.mk: Wait for terminals to be started before running QEMU
At the moment we start several terminals, and then sleep for a second
before running a QEMU instance which will connect to processes run
in those terminals. This will fail if the terminals are displayed on
a slow remote X11 connection, because the terminals won't have
completed startup before the one second delay elapses.

Replace the delay with a loop that checks whether there's anything
listening on the relevant TCP ports yet.

Fixes https://github.com/OP-TEE/build/issues/20

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Cedric Chaumont <cedric.chaumont@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org>