hikey: Add sleep to prevent umount failure
Sometimes it happens that the umount fails saying that a device/resource
is busy when trying umount the device after the filesystem has been
created. We don't know how and why this happens, but in the meantime
while waiting for a proper solution we have added a sleep that seems to
solve the problem.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>