Fix the errors present in QEMU instructions-a64.h
When compiling OP-TEE for QEMU with a recent version of GCC it
causes build errors `-Werror=unused-variable`.

By adding `--extra-cflags="-Wno-error"` to the makefile it safely
supresses the errors and allows OP-TEE to be compiled and built. This
error is also present in the `default_stable.xml` configuration with the
repo tool of the project.

See [OPTEE_OS issue 550](https://github.com/OP-TEE/optee_os/issues/550).