HiKey: Add recovery target and instructions
To make it a bit easier to understand how to do a recovery and how to
flash HiKey, we have added a new make target for recovery which tells
how to configure the jumpers and the initiates the recovery itself. In
addition to this we have changed the "flash" make target so that it
does the same as stated in the instructions on the 96boards "Getting
started" page.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (HiKey)
Reviewed-by: Victor Chong <victor.chong@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>