qemu: Ability to switch Python version for QEMU
On some distros /usr/bin/python points to Python 3.x and changing that
to point to Python 2.x might not be an easy thing to do. By adding the
variable PYTHON_PATH one can easily change the python version to be used
with QEMU (QEMU only supports Python 2.x).

For a distro where Python 3.x is the default one could/should call make
like this:

 $ PYTHON_PATH=/usr/bin/python2 make

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>