qemu: fix linux kernel config regarding SDP support
The first 1Mbyte of the secure RAM is reserved to the ARM-TF on
the qemu ARMv8 setup. To support both QEMU ARMv7 and QEMU ARMv8
with the same kernel configuration that includes a static definition
of the SDP test memory pool, move the SDP pool to the end of the
QEMU secure RAM. SDP pool size is increased to 4MByte.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>