README: Remove `run` target
Only QEMU, FVP and MTK8173 have a `run` target but all *.md files
refer to README Section 7.5 which calls for `make all run`. Remove
this target and add it to mtk8173.md instead. qemu.md and fvp.md
already mentions the run target so don't need to be added again.

Fixes: https://github.com/OP-TEE/build/issues/146
Signed-off-by: Victor Chong <victor.chong@linaro.org>