travis: Always use the branch from the committer
Regardless if it is a push or a pull request, Travis will use the
committers repository and branch. For developers this means his own
branch will be used when pushing to his own gits and when making a pull
request. When the pull request is merged, which is seen as normal push
to the upstream repository, Travis will once more build according to the
definition, but this time use the upstream repository and the tip branch
instead.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Igor Opaniuk <igor.opaniuk@linaro.org>