common.mk: allow user to override launch-terminal in make run
Currently the code is hard coded to launch netcat via either xterm or
gnome-terminal. This is a little inflexible for people who spend a lot
of time in terminals. With the following you can now do:

  env QEMU_PATH=~/lsrc/qemu/qemu.git/ \
      LAUNCH_TERMINAL="tmux split-window" \
      make -j run

To bring your nc instances up in new tmux windows. The one downside is
titles aren't set as that wouldn't be maximally flexible for whatever
your launch command is.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>