fvp: Introduce grub to the FPV setup
As a first step to update EDK2 to running latest  we are introducing
grub which will be a needed component when booting UEFI without
semihosting.

This particular patch will just bulid grub, it will still not be used
when running FVP (that is enabled in later patches in this series).

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Victor Chong <victor.chong@linaro.org>