fvp: Build and use an UEFI boot image
We are switching from using the old semihosting based way of loading and
running OP-TEE, kernel, root fs etc on FVP to instead use a proper boot
image similar to how it is working on real devices. A difference here
compared to the past is that we are using Grub here to load the kernel.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (FVP)
Reviewed-by: Victor Chong <victor.chong@linaro.org>