edk2: Update QEMU v8 and FVP builds
QEMU v8 and FVP are both now working with upstream edk2 and
edk2-platforms. QEMU v8's dsc file is located in the edk2 folder whilst
the dsc for FVP is in the edk2-platforms folder. Due to this we had to
export the WORKSPACE variable differently otherwise we would get either
compiler errors or we would get the "Build" folder created on the top
level (something that we do not want).

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (QEMU v8, FVP)
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>