qemu_v8: fix missing edk2 'cleanall' target
The edk2-clean-common target in common.mk uses a cleanall target, which does
not exist in the makefile used by edk2-call in qemu_v8.mk. Adding a specific
target for edk2-clean in qemu_v8.mk that uses 'clean' instead of 'cleanall'
allows the clean target to complete successfully.

Signed-off-by: Joel Anderson <jeander@vt.edu>