travis: rename the github remote
With the recent QEMU v7 changes in manifest.git:
  918f04199c ("qemu: general cleanup and remote changes")
where the remote "github" was introduced, we introduced a regression for
the QEMU v7 Travis builds. The remote name "github" has been used in
Travis when setting up a remote for the build.git itself. By renaming
the remote we do not get any naming collisions and the Travis build
script will be prepared for similar changes in manifest.git for the
other platforms.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>