qemu: make check: create symlink to rootfs.cpio.gz
Now that Arm Trusted Firmware is used as the secure boot loader
and bios_qemu_tz_arm is used as the non-secure bootloader, all binaries
are expected to be found in the $(BINARIES) directory. Symbolic links
are created by various targets in the qemu.mk.
The 'check' target should create a link to rootfs.cpio.gz, similar to
the 'run' target.

Fixes: 46ddff2b339b9 ("qemu: boot from arm-tf")
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Joakim Bech <joakim.bech@linaro.org>