common: remove deprecated rootfs targets
Targets update_rootfs-common, update_rootfs-clean-common and
filelist-tee-common are not used by any of the supported
platform and their content is not accurate regarding OP-TEE
packages content (init.d.optee, tee-supplicant install path).

This change removes this make targets.

This change also remove the init.d.optee init script which is
no more embedded in the filesystem of the supported platforms.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>