hikey960: enable CFG_SECURE_DATA_PATH by default
Now that our HiKey960 build uses our common Linux branch [1], we no
longer need to force CFG_SECURE_DATA_PATH=n in the main Makefile (the
platform branch we were using previously was lacking the ION unmapped
heap code, which is why SDP had to be disabled).

Let OP-TEE build with the default value for CFG_SECURE_DATA_PATH, which
is 'y' for HiKey960, and configure the Linux kernel accordingly.

[1] https://github.com/linaro-swg/linux branch optee. This branch has
been used for the HiKey960 build since OP-TEE 3.4.0, the branch was
based on upstream kernel v4.19.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>