common.mk: buildroot: append BR2_ variables to config automatically
To set additional BR2_ variables, the user currently needs to edit
common.mk and modify the buildroot recipe. It is inconvenient,
especially since a line with a cumbersome syntax has to be added. One
would expect to be able to specify the desired value on the make
command line or directly in the Makefile. For instance, to enable
debug mode: make BR2_ENABLE_DEBUG=y.

This patch updates the buildroot recipe to enumerate all the Makefile
variables that start with BR2_, and each such variable is automatically
added to ../out-br/extra.conf.

As a result the existing BR2_ variables can be set in a simpler way.

Tested on QEMU, results in the same ../out-br/.config except for the
line order.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Suggested-by: Heino Madsen <heino.madsen@huawei.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>