br-ext: optee_test: enumerate build targets via shell instead of make
Depending on the size of the string in the variable
TARGET_CONFIGURE_OPTS and number of tests, the size of the
string in OPTEE_TEST_BUILD_TAS produced by evaluation of
the $(foreach...) statement may exceed the maximum length for
execvp() calls from make (128KiB). Delaying the enumeration of
targets by passing a loop-statement for the shell expands the
variables only once and keep the length of the string independent
of the number of tests.

Signed-off-by: Markus S. Wamser <markus.wamser@mixed-mode.de>
Acked-by: Jerome Forissier <jerome@forissier.org>