GitHub: sync GitHub templates and actions from optee_os
The GitHub templates and actions are useful to the "build" git also,
therefore recursively sync up (copy) the .github folder from optee_os.

Note, that here we have also done some changes to the GitHub actions. In
optee_os we have two different actions files, one for issues and one for
pull requests. It seems like it is possible to combine them into a
single action file as shown in the actions examples [1]. Another change
is that we also exclude marking issues stale if they already have the
label "bug". Note that we're only excluding issues, since PR's should
not have the "bug" label to start with.

Link: [1] https://github.com/actions/stale/blob/master/README.md

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>