rpi3: avoid buildroot architecture checks for boot
Buildroot checks the ELF header for each ELF it installs and compares
the MACHINE field with the architecture it is building for. If there
is a mismatch the build fails. The Raspberry 3 build install several ELF's
to the boot partition of type "Broadcom VideoCore III processor" and
not Aarch64 as expected by Buildroot.

This patch adds directory 'boot' to the list of directories excluded
from architecture checking.

Signed-off-by: GlovePuppet <touched.by.his.noodley.appendage@gmail.com>
Acked-by: Joakim Bech <joakim.bech@linaro.org>