toolchain: update to GCC 9.2
Updates the toolchain to Arm's GCC 9.2 [1]. Note the following
difference compared to the previous release (8.3), which is documented
in the release notes:

 "Changed toolchain naming convention to match standard target triplet
  naming convention, with vendor name being none."

For example, arm-linux-gnueabihf-gcc is now arm-none-linux-gnueabihf-gcc
and aarch64-linux-gnu-gcc is now aarch64-none-linux-gnu-gcc.

This renaming seems pretty much specific to the Arm build of GCC, and
the latest Debian and Ubuntu distributions provide GCC 9.x or 10.x
cross-compilers with the usual naming convention.

Therefore, and to limit the impact of the name change, let's simply
create symbolic links after the toolchain packages are extracted.

Link: [1] https://developer.arm.com/tools-and-software/open-source-software/developer-tools/gnu-toolchain/gnu-a/downloads
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Tested-by: Jerome Forissier <jerome@forissier.org> (QEMU, QEMUv8, HiKey, HiKey960)
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Tested-by: Etienne Carriere <etienne.carriere@linaro.org> (stm32mp1)