Introduce BR2_PACKAGE_OPTEE_TEST_EXT_WITH_CXX_TESTS
Conditionally enable TA C++ tests in the optee_test_ext Buildroot package.
Some toolchains may not support this feature [1] so it is best to let
the user decide.

common.mk enables the tests by default when $(COMPILER) is not clang,
assuming that the default GCC toolchain from toolchain.mk is used (which
works fine with the C++ tests). The command line option is
'make WITH_CXX_TESTS=y|n'.

Link: https://github.com/OP-TEE/optee_test/issues/458
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>