rpi3: recommend 64MB for the boot partition
In some cases depending on configurations etc 32MB is a bit tight for
the boot partition. By increasing it to 64MB you are in a better
position for not hitting the limit. Likewise you can store additional
kernel versions in the boot partition which can be useful when debugging
issues showing up in one kernel build, but not the other. Because of
this the help text should state 64MB instead of 32MB.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>