common.mk: qemu: do not build docs
Add the --disable-docs option to QEMU_CONFIGURE_PARAMS_COMMON so that no
documentation is generated during the QEMU build. Documentation is not
needed and it may fail to build in some environments. For instance on
Ubuntu 18.04.5 LTS:

 Warning, treated as error:
 docs/qemu-option-trace.rst.inc:4:Malformed option description '[enable=]PATTERN', should look like "opt", "-opt args", "--opt args", "/opt args" or "+opt args"
 make[1]: *** [.docs_built_system_qemu.1_docs_built_system_qemu-block-drivers.7_docs_built_system_qemu-cpu-models.7.sentinel.] Error 2

The error is caused by sphinx-build:

 $ sphinx-build --version
 sphinx-build 3.2.1

The version that comes with Ubuntu 20.04.1 LTS works fine:

 $ sphinx-build --version
 sphinx-build 1.8.5

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>