Add support for NXP i.MX8MQ EVK
Adds support for the NXP® i.MX 8MQuad Evaluation Kit (EVK) using the
same standard buildroot setup as for the other platforms here.

The flash-image target (part of the all target) produces a out/boot.img
which can be copied to a SD card using the command:
dd if=out/boot.img of=/dev/sdX bs=1M conv=fsync

Where 'X' has to be adjusted to the device of the SD card when connected
to the host system.

     ****** Please be careful to choose the correct device ******
     ****** or there's a risk of damaging the host system. ******

Acked-by: Joakim Bech <joakim.bech@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>