linux-defconfig: explicit cross compile flag usage
Current linux config file generation does not add the cross compilation
flag, leading potentially to compilation errors. Not sure how a proper config
file is generated with current implementation but issues are observerd
specially when buildling the linux image with a single thread.

This commit ensures that both ARCH and CROSS_COMPILE flags are set when
using the merge_config script.

Signed-off-by: Ibai Erkiaga <ibai.erkiaga-elorza@xilinx.com>
Reviewed-by: Jerome Forissier <jerome@forissier.org>