stm32mp1: build tf-a images without FIP support
Explicitly enable platform config switch STM32MP_USE_STM32IMAGE to
enforce TF-A legacy boot of OP-TEE and U-Boot images in several storage
partitions. Platform stm32mp1 changed in TF-A v2.6 in that the default
configuration do enable use of a FIP image.

Upgrading platform build to use FIP is out of the scope of this change.

Acked-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>