toolchain: Add optimizations for aarch64 host
Lets switch over to using aarch64 hosted cross-compiler for aarch32
rather than building toochain for aarch32 from source which is very
cumbersome. But in case of native build for aarch64 on aarch64 host we
are left with no choice but to use buildroot toolchain as Arm doesn't
provide binary toolchain releases for aarch64 native compiler. Also,
we can't rely on distribution toolchain as buildroot doesn't support it
(refer here [1] for details).

Link: [1] https://buildroot.org/downloads/manual/manual.html#_cross_compilation_toolchain
Signed-off-by: Sumit Garg <sumit.garg@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>