br-ext: opensc: fix dependencies
When BR2_PACKAGE_OPENSSL=y, OPENSC_DEPENDENCIES is overwritten with
'openssl' which is wrong. 'openssl' should be appended instead,
otherwise the dependency on 'pcsc-lite' is lost.

This issue was found when building with BR2_PER_PACKAGE_DIRECTORIES=y
which triggers a parallel build. It looks like the non-parallel build
works by chance (a matter of build order probably).

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>