fvp: add Trusted Services support
Add support for building Secure Partitions implemented by the Trusted
Services project. The patch contains helper functions to invoke the TS
buildsystem in order to build SPs and include them into the OP-TEE bin.
It also adds a new top level Makefile which configures OP-TEE and TF-A
for the S-EL1 SPMC scenario on FVP platform.

Link: https://www.trustedfirmware.org/projects/trusted-services/
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Gyorgy Szing <gyorgy.szing@arm.com>
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>