fvp: separate Measured Boot and fTPM config
The Measured Boot option not only enables the relevant config options
for TF-A and optee_os, but also building the fTPM TA and its Linux
driver. This is unnecessary when the consumer of the Event Log is not
fTPM but e.g. the PSA Attestation SP (as introduced in the next commit).

This commit modifies the MEASURED_BOOT option to only set the TF-A and
optee_os config, and adds the MEASURED_BOOT_FTPM option for the fTPM
related components. To keep backwards compatibility if MEASURED_BOOT is
selected, by default MEASURED_BOOT_FTPM is enabled too.

Reviewed-by: Gyorgy Szing <gyorgy.szing@arm.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>