buildroot: select util-linux-libs, libuuid and pkgconf
In order to use login methods `TEEC_LOGIN_USER` and `TEEC_LOGIN_GROUP`,
a user space application must initialize the tokens with the SHA1
hashed UUID of the user or group.

Select host-pkgconf, util-linux and libuuid in order to use `libuuid`
and add supporting libraries to optee_client for this type of ACL based
login initialization.

Signed-off-by: Eero Aaltonen <eero.aaltonen@vaisala.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>