Add support for ROCK 4 (aka Rockpi4)
Add rockpi4.mk and as well as U-Boot and Linux kernel config fragments
plus one TF-A patch (fix console baudrate) and one U-Boot patch (enable
OP-TEE in the DT).

Similar to other platforms such as HiKey, typical build commands are:

  $ make toolchains
  $ make -j$(nproc)
  $ make flash

The bootable image is out/rockpi4.img (128 MB). "make flash" transfers
it to the onboard eMMC, but it may also be written to a micro SD card.
Note however that the board tries to boot from eMMC first, so once "make
flash" has been used, the eMMC has to be cleared to be able to boot from
micro SD again. This can be done with "make nuke-emmc".

Tested on Rockpi4B.

Link: https://wiki.radxa.com/Rock4
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>