trusted-keys.exp: fix error detection
The check_keyctl_result() procedure doesn't deal with all error cases
properly. For example the following error is not reported:

 # keyctl add trusted kmk "new 32" @u
 [    7.293607] trusted-key-tee optee-ta-f04a0fe7-1f5d-4b9b-abf7-619b85b4ce8c: key shm register failed
 add_key: Invalid argument
 #

To make things simpler, rely on the exit status of the commands sent to
the guest. The || shell construct is used to trigger the display of an
error message ("FAILED"). To avoid having the message in the command
themselves, which would make the parsing more complex, a shell function
is introduced ("fail"). Then all the check_keyctl_result() procedure
needs to do is wait for the next prompt ("# "), while dealing with
errors (the "FAILED" word appearing anywhere), key ID (parsing is
unchanged), and timeouts (which I have reduced to 5 seconds, more than
enough for any operation in this test suite).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Sumit Garg <sumit.garg@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>