buildroot: remove opensc external package
Buildroot supports opensc since commit 2022.05-701-g8aaa7ecbce, and the
new package conflicts with our local copy in br-ext/. Therefore, remove
the latter.

Note that with this change it is not possible to configure opensc
to use a custom PKCS#11 provider at build time. So in order to
use OP-TEE with pkcs11-tool, the '--module /usr/lib/libckteec.so.0'
option has to be added when invoking the command.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>