fvp: fix TF-A parameter passing issue
Build fails if MEASURED_BOOT=n is passed on the command line to
fvp-psa-sp.mk due to colliding variable names between this repo
and TF-A makefiles.  Both projects use "boolean type", but the
implementation is incompatible. OP-TEE uses "y/n" and TF-A uses
"1/0" as the true/false value.
Make will pass variables set on the command-line down to all
sub-make invocations, and thus if the value is not converted
TF-A build will fail.

This change fixes the issue by adding a missing value conversion.

Signed-off-by: Gyorgy <gyorgy.szing@arm.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>