ts: Generate sp_layout.json after SP builds
TF-A requires to have the SPs in the same order in the FIP as it is
defined in ARM_BL2_SP_LIST_DTS. The sp_layout.json controls the order
of SPs in the FIP package. Previously the individual JSON fragments were
appended to the sp_layout.json file at the end of the SP build. The
final order of SP definitions in the sp_layout.json depended on the SP
build order which was non-deterministic because of the parallel make
jobs. To resolve this issue the JSON merging step is now done after all
SPs have been built. The order will follow the order of build-sp calls
in the fvp-psa-sp.mk which matches the order of SPs in
ARM_BL2_SP_LIST_DTS.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>