qemu_v8: add uImage target
"make linux" builds the kernel but is not sufficient to prepare all
files for "make run-only". On the other hand, "make run" does a bit
too many things to be used comfortably when working on the kernel.

Therefore, add target uImage which generates the kernel image ready
to be used by U-Boot. With that, kernel development and testing can
be done with the following sequence:

 1. make
 2. Edit kernel sources
 3. make -j$(nproc) uImage
 4. make run-only
 5. Go to 2 -- repeat as needed

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>