rpi3: fix dependency issue when building
The Raspberry Pi 3 build appears to complete with no errors from time to
time, but it is still an incomplete build because the image for the SD
card for example has not been built. This is most likely due to a
dependency error in the Makefile. After reviewing the dependencies, it
appears that at least two of them were incorrect or absent.

- Add 'u-boot' instead of 'u-boot-env' as a dependency to TF-A.
- Add 'linux' and 'u-boot-env' as a dependency to the boot filesystem.

Fixes: https://github.com/OP-TEE/optee_os/issues/6442
Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>