toolchains.mk, get_clang.sh: upgrade to Clang 18.1.6
Update the Clang version from 12.0.0 to 18.1.6. Unfortunately, newer
versions of the LLVM toolchain do not make available the pre-built
compiler-rt packages for arm64 and armhf. Therefore, the way we obtain
the toolchain is changed. I created a project with a Makefile and a
Dockerfile [1] which allows to build Clang from scratch with the proper
compiler-rt libraries for OP-TEE cross-compilation. Nothing really
special, yet not a configuration that is availale pre-built in the
major Linux distributions (Unbuntu). The problem is that on an x84_64
system, there is no way to install the compiler-rt builtins for a
different architecture (libclang_rt.builtins-{aarch64,armhf}.a).

The new method to install Clang is thererfore to pull a pre-built
image (called jforissier/clang-<version>-<arch>) from the Docker
Hub, then simply extract the files contained inside, which is what
the modified get_clang.sh does.

Link: https://github.com/jforissier/docker_clang [1]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>